REFACTOR[rutas, screens, search]: refactorizo un poco rutas, y creo nueva ruta de search
