+++ src/app/product/[id]/page.tsx
@@ -2,9 +2,5 @@ import { ProductDetailScreen } from "@/components/screens";
 import React from "react";
 
 export default function Product() {
-  return (
-    <div>
-      <ProductDetailScreen />
-    </div>
-  );
+  return <ProductDetailScreen />;
 }
