+++ src/app/search/page.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import { SearchScreen } from '@/components/screens'
+
+export default function page() {
+    return <SearchScreen />
+}
