+++ src/components/screens/searchScreen/SearchScreen.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export const SearchScreen = () => {
+    return (
+        <div className='mt-40 bg-slate-600'>
+            tienda de productos SCREEN
+
+        </div>
+    )
+}
